WIP: Add tox config
This is an attempt at using tox again, ideally instead of invoke.
Last attempt was frustrating so I decided to stick to using invoke
but it's not particularly standard and I am seeing more and
more Python packages standardasing on tox.

This commit adds a tox config for most useful things and refactors
requirements, configs accordingly so it's usable locally (not
used for CI).