Make it so default resolvers are defined through the Schema class
As field resolvers were defined through the Schema class it
seemed consistent to do the same for default resolvers.

This adds support for this in ResolverMap and Schema (still
forwarding to the relevant types) through `register_default_resolver`
and @resolver("Type.*") for type level defaults as well
as Schema.default_resolver for the global default.

This also drops support for the default_resolver argument
in various execution related functions so it is only defined
alongside the schema.